refactor: group LoginController dependencies in a single file
issue: #3
