+++ src/presentation/controllers/login/login.spec.ts
@@ -1,7 +1,6 @@
-import { Authentication } from '../../../domain/usecases/authentication'
+import { EmailValidator, HttpRequest,Authentication } from './login-protocols'
 import { InvalidParamError, MissingParamError } from '../../errors'
 import { badRequest, serverError, unaunthorized } from '../../helpers/http-helper'
-import { EmailValidator, HttpRequest } from '../signup/signup-protocols'
 import { LoginController } from './login'
 
 const MakeEmailValidator = (): EmailValidator => {
